requirements.txt
aioitertools==0.12.0
annotated-types==0.7.0
attrs==25.1.0
bcrypt==4.2.1
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dnspython==2.7.0
emmet-core==0.84.6rc4
fonttools==4.56.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
latexcodec==3.0.0
maggma==0.71.4
matplotlib==3.10.0
mongomock==4.3.0
monty==2025.1.9
mp-api==0.45.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.26.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.15
packaging==24.2
palettable==3.3.3
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
plotly==6.0.0
pyaml_env==1.2.2
pybtex==0.24.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydash==8.0.5
pymatgen==2025.1.24
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.2
scipy==1.15.1
sentinels==1.0.0
six==1.17.0
smart-open==7.1.0
spglib==2.5.0
sshtunnel==0.4.0
sympy==1.13.3
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uncertainties==3.2.2
urllib3==2.3.0
wrapt==1.17.2
